Enable Map / Set support by default
