+++ __tests__/plugins.js
@@ -1,11 +1,5 @@
 import {produce, produceWithPatches, applyPatches} from "../src/immer"
 
-test("error when using Maps", () => {
-	expect(() => {
-		produce(new Map(), function() {})
-	}).toThrowErrorMatchingSnapshot()
-})
-
 test("error when using patches - 1", () => {
 	expect(() => {
 		produce(
