+++ src/immer.ts
@@ -103,4 +103,3 @@ export function castImmutable<T>(value: T): Immutable<T> {
 export {Immer}
 
 export {enablePatches} from "./plugins/patches"
-export {enableMapSet} from "./plugins/mapset"
