+++ src/internal.ts
@@ -9,3 +9,4 @@ export * from "./core/finalize"
 export * from "./core/proxy"
 export * from "./core/immerClass"
 export * from "./core/current"
+export * from "./core/mapset"
